+++ cmd/isosim/version.go
@@ -1,4 +1,4 @@
 package main
 
 var version = "0.8.0"
-var build = "d01aad2c"
+var build = "741453a7"
