+++ internal/iso/server/iso_def.go
@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"isosim/internal/db"
-	"isosim/internal/services/v0/data"
+	"isosim/internal/services/data"
 
 	"sync"
 )
