+++ internal/services/handlers/misc_handler.go
@@ -1,13 +1,10 @@
-package misc
+package handlers
 
 import (
-	"bytes"
 	"encoding/hex"
-	"encoding/json"
 	"github.com/rkbalgi/libiso/hsm"
 	"github.com/rkbalgi/libiso/net"
 	log "github.com/sirupsen/logrus"
-	"isosim/internal/db"
 	"isosim/internal/iso"
 
 	"net/http"
@@ -25,62 +22,7 @@ func init() {
 
 }
 
-const defaultFormat = "json"
-
 func AddMiscHandlers() {
-
-	http.HandleFunc("/iso/v1/websim/msg_hist/last_n", func(rw http.ResponseWriter, req *http.Request) {
-
-		rw.Header().Add("Access-Control-Allow-Origin", "http://localhost:3000")
-
-		if err := req.ParseForm(); err != nil {
-			_, _ = rw.Write([]byte(err.Error()))
-			rw.WriteHeader(http.StatusBadRequest)
-			return
-		}
-
-		format := defaultFormat
-
-		msgId, _ := strconv.Atoi(req.Form.Get("msg_id"))
-		specId, _ := strconv.Atoi(req.Form.Get("spec_id"))
-		count, _ := strconv.Atoi(req.Form.Get("count"))
-		format = req.Form.Get("format") //can be json or html
-
-		if format == "" {
-			format = defaultFormat
-		}
-
-		if res, err := db.ReadLast(specId, msgId, count); err != nil {
-			_, _ = rw.Write([]byte(err.Error()))
-			rw.WriteHeader(http.StatusBadRequest)
-		} else {
-
-			if format == defaultFormat {
-				jsonResp, _ := json.Marshal(res)
-				_, _ = rw.Write(jsonResp)
-				return
-			}
-
-			buf := bytes.Buffer{}
-			if len(res) > 0 {
-				buf.Write([]byte(`<html><body>`))
-			} else {
-				rw.Write([]byte("No records found.."))
-				return
-			}
-
-			for _, tmp := range res {
-				buf.Write([]byte(`<div style="color:blue; background-color:azure; border-style:ridge;">`))
-				buf.Write([]byte(tmp))
-				buf.Write([]byte("</div></hr>"))
-			}
-
-			buf.Write([]byte(`</body></html>`))
-			rw.Header().Add("Content-Type", "text/html")
-			_, _ = rw.Write(buf.Bytes())
-		}
-
-	})
 
 	http.HandleFunc("/iso/misc", func(rw http.ResponseWriter, req *http.Request) {
 
@@ -201,11 +143,5 @@ func AddMiscHandlers() {
 		}
 
 	})
-
-}
-
-func sendError(rw http.ResponseWriter, err string) {
-	rw.WriteHeader(500)
-	rw.Write([]byte(err))
 
 }
