+++ internal/services/websim/service.go
@@ -12,7 +12,7 @@ import (
 	log "github.com/sirupsen/logrus"
 	"isosim/internal/db"
 	"isosim/internal/iso"
-	"isosim/internal/services/v0/data"
+	"isosim/internal/services/data"
 	"net"
 	"sort"
 	"strconv"
